requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-text-splitters==1.1.2
google-genai==1.73.1
pypdf==6.10.2
sentence-transformers==5.4.1
chromadb==1.5.8
tabula-py==2.10.0
pandas==2.3.3
jpype1==1.7.0
streamlit==1.56.0

